// src/core/types/physical-characteristic-type.ts
export enum PhysicalCharacteristicType {
  WEIGHT = "Weight",
  HEIGHT = "Height",
  LENGTH = "Length",
  CHEST_GIRTH = "Chest girth",
  NECK_GIRTH = "Neck girth",
  WINGSPAN = "Wingspan",
  TAIL_LENGTH = "Tail Length",
  SHELL_LENGTH = "Shell Length",
  SHELL_WIDTH = "Shell Width",
  BODY_TEMPERATURE = "Body Temperature",
  BODY_CONDITION = "Body Condition Score",
}