// src/core/types/treatment-type.ts
export enum TreatmentType {
  VACCINE = "Vaccine",
  ANTIBIOTIC = "Antibiotic",
  ANTIPARASITIC = "Antiparasitic",
  VITAMINS = "Vitamins",
  SUPPLEMENTS = "Supplements",
  SURGERY = "Surgery",
  HORMONAL = "Hormonal Therapy",
  THERAPY = "Therapy",
  REHABILITATION = "Rehabilitation",
  PAIN_RELIEF = "Pain Relief",
  DIETARY = "Dietary / Veterinary Diet",
  OTHER = "Other",
}